opea-comps/mega-service/requirements.txt
opea-comps
fastapi
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi